+++ packages/app/src/lib/components/content/thread/ChatArea.svelte
