+++ packages/app/src/lib/components/content/thread/ChatInput.svelte
