+++ packages/app/src/lib/components/content/thread/message/ChatMessage.svelte
