+++ packages/app/src/lib/components/content/thread/message/MessageReactions.svelte
