+++ packages/app/src/lib/components/content/thread/message/MessageThreadBadge.svelte
