+++ packages/app/src/lib/components/sidebars/SidebarObject.svelte
