+++ packages/app/src/lib/components/sidebars/SidebarObjectList.svelte
