+++ packages/app/src/lib/components/spaces/SpaceAvatar.svelte
